add untested .mtl parser
