+++ examples/bunny.rs
@@ -14,11 +14,12 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     scene.add(
         Mesh::new(
-            obj::obj_from_url(BUNNY_URL)?,
+            parser::obj_from_url(BUNNY_URL)?,
             Material::transparent(
                 Texture::Solid(srgb_to_linear(0, 255, 0)),
+                0.1,
                 1.5,
-                0.1),
+            ),
         )
         .scale(0.3, 0.3, 0.3)
         .translate(0.0, -0.65, -1.5),
